feat: modify type and error

+++ dubeng-user/src/main/java/com/ssafy/dubenguser/entity/DubKing.java
@@ -17,12 +17,12 @@ public class DubKing extends Time{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @ManyToOne(cascade = CascadeType.MERGE, targetEntity = User.class, fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
-    private long totalVote;
+    private Long totalVote;
     @UpdateTimestamp
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSXXX")
     private ZonedDateTime endDate;
